autonomous_dj/observability.py
#!/usr/bin/env python3
"""
Observability Module - Tracing and Monitoring Integration
==========================================================

Integrates datapizza-ai ContextTracing with OpenTelemetry for production monitoring.

Features:
- LLM operation tracking (token usage, cost, latency)
- DJ operation tracking (browser navigation, track loading, mixing)
- Automatic span creation and collection
- Export to Grafana/Datadog/Jaeger via OTLP

Usage:
    from autonomous_dj.observability import trace_operation, enable_detailed_tracing

    # Wrap any DJ operation
    with trace_operation("load_track_to_deck_a"):
        load_track(deck='A', track_path=...)

    # Enable detailed I/O tracing
    enable_detailed_tracing()

Author: DJ Fiore AI System
Version: 1.0
Created: 2025-10-19 (Phase 3 - datapizza-ai integration)
"""

import os
from typing import Optional, Dict, Any
from contextlib import contextmanager
from datapizza.tracing import ContextTracing

# Initialize tracing system
_tracer = ContextTracing()


def enable_detailed_tracing():
    """
    Enable detailed I/O tracing for LLM operations

    Sets environment variable to trace all client I/O operations.
    Useful for debugging but increases logging verbosity.
    """
    os.environ["DATAPIZZA_TRACE_CLIENT_IO"] = "true"
    print("[OBSERVABILITY] Detailed I/O tracing enabled")


def disable_detailed_tracing():
    """Disable detailed I/O tracing"""
    os.environ["DATAPIZZA_TRACE_CLIENT_IO"] = "false"
    print("[OBSERVABILITY] Detailed I/O tracing disabled")


@contextmanager
def trace_operation(operation_name: str, metadata: Optional[Dict[str, Any]] = None):
    """
    Context manager for tracing DJ operations

    Args:
        operation_name: Name of the operation (e.g., "load_track", "execute_transition")
        metadata: Optional metadata to attach to the trace

    Usage:
        with trace_operation("browser_navigate", {"target": "Dub", "scrolls": 4}):
            navigate_to_folder("Dub", scroll_count=4)

        # Trace summary automatically displayed:
        # ╭─ Trace Summary of browser_navigate ──────────────╮
        # │ Total Spans: 3                                   │
        # │ Duration: 13.1s                                  │
        # ╰──────────────────────────────────────────────────╯
    """
    with _tracer.trace(operation_name) as trace:
        # Add metadata if provided
        if metadata:
            # Store metadata in trace context (if supported)
            trace.metadata = metadata

        try:
            yield trace
        except Exception as e:
            # Log exception in trace
            print(f"[OBSERVABILITY] Operation '{operation_name}' failed: {e}")
            raise


def trace_llm_decision(llm_func):
    """
    Decorator for tracing LLM decision functions

    Automatically tracks:
    - Prompt tokens
    - Completion tokens
    - Total tokens
    - Execution time
    - Model used
    - Cost estimate

    Usage:
        @trace_llm_decision
        async def get_llm_decision(input_query, state):
            # ... LLM call ...
            return decision
    """
    async def wrapper(*args, **kwargs):
        # Extract operation name from function
        operation_name = f"llm_{llm_func.__name__}"

        with trace_operation(operation_name):
            result = await llm_func(*args, **kwargs)
            return result

    return wrapper


def trace_autonomous_workflow(workflow_func):
    """
    Decorator for tracing complete autonomous workflows

    Tracks end-to-end operations:
    - Track discovery
    - Browser navigation
    - Deck loading
    - Playback start

    Usage:
        @trace_autonomous_workflow
        def run_complete_workflow(target_folder, scroll_count):
            # ... autonomous workflow ...
            pass
    """
    def wrapper(*args, **kwargs):
        # Extract operation name
        operation_name = f"workflow_{workflow_func.__name__}"

        with trace_operation(operation_name):
            result = workflow_func(*args, **kwargs)
            return result

    return wrapper


class PerformanceMonitor:
    """
    Performance monitoring class for DJ operations

    Tracks operation latency and ensures <10ms MIDI latency requirement.
    """

    def __init__(self):
        self.latency_threshold_ms = 10  # <10ms MIDI latency target
        self.warnings = []

    @contextmanager
    def monitor_latency(self, operation_name: str):
        """
        Monitor operation latency and warn if exceeds threshold

        Usage:
            monitor = PerformanceMonitor()
            with monitor.monitor_latency("midi_command"):
                send_midi_cc(cc=47, value=127)
        """
        import time

        start = time.time()
        try:
            yield
        finally:
            duration_ms = (time.time() - start) * 1000

            if duration_ms > self.latency_threshold_ms:
                warning = f"[PERFORMANCE WARNING] {operation_name} took {duration_ms:.2f}ms (>{self.latency_threshold_ms}ms)"
                self.warnings.append(warning)
                print(warning)


# Global performance monitor instance
performance_monitor = PerformanceMonitor()


def get_trace_summary(trace):
    """
    Get human-readable trace summary

    Args:
        trace: Trace object from trace_operation context manager

    Returns:
        Dict with span count, duration, token usage
    """
    spans = trace.get_spans() if hasattr(trace, 'get_spans') else []

    summary = {
        'span_count': len(spans),
        'operation': trace.name if hasattr(trace, 'name') else 'unknown',
        # Duration and token info extracted from spans if available
    }

    return summary


# Example usage documentation
if __name__ == "__main__":
    print("=" * 60)
    print("OBSERVABILITY MODULE - Example Usage")
    print("=" * 60)
    print()

    # Example 1: Trace a simple operation
    print("Example 1: Trace browser navigation")
    with trace_operation("browser_navigate", metadata={"target": "Dub"}):
        import time
        time.sleep(0.1)  # Simulate navigation
        print("  [SIMULATED] Navigated to Dub folder")

    print()

    # Example 2: Enable detailed tracing
    print("Example 2: Enable detailed I/O tracing")
    enable_detailed_tracing()
    print("  Status: DATAPIZZA_TRACE_CLIENT_IO =", os.environ.get("DATAPIZZA_TRACE_CLIENT_IO"))

    print()

    # Example 3: Performance monitoring
    print("Example 3: Monitor MIDI latency")
    with performance_monitor.monitor_latency("test_midi_command"):
        import time
        time.sleep(0.005)  # 5ms - within threshold

    print()

    with performance_monitor.monitor_latency("slow_midi_command"):
        import time
        time.sleep(0.015)  # 15ms - exceeds threshold!

    print()
    print("=" * 60)
    print("Observability module ready for production use")
    print("=" * 60)
